Changed the structure, moved functions from board.java to TicTacToe.java
