+++ src/main/java/is/ru/tictactoe/Board.java
@@ -13,77 +13,6 @@ public class Board {
 			 }
 		 }
 	 }
-	 public static boolean isWinner(int row, int col){
-		 boolean winner = false;
-		 char player = board[row][col];
-		 int count = 0;
-		 
-		 //athugum larett
-		 for(int i = 0; i < SIZE; i++){
-			 if(board[i][col] == player)
-				 count++;
-		 }
-		 winner = win(count);
-		 
-		 //athugum lodrett
-		 for(int j = 0; j < SIZE; j++){
-			 if(board[row][j] == player)
-				 count++;
-		 }
-		 winner = win(count);
-		 
-		 //athugum a ska fra vinstri
-		 for(int i = 0; i < SIZE; i++){
-			 if(board[i][i] == player)
-				 count++;
-		 }
-		 winner = win(count);
-		 
-		 //athugum a ska fra haegri
-		 for(int j = 0; j < SIZE; j++){
-			 if(board[j][2-j] == player)
-				 count++;
-		 }
-		 winner = win(count);
-		 
-		 return winner;
-		 
-	 }
-	 public static boolean win(int count){
-		 boolean win = false;
-		 if(count == SIZE) {
-			 win = true;
-		 }
-		 count = 0;
-		 return win;
-	 }
-
-	 public static boolean full(){
-	 	for(int i = 0; i < SIZE; i++){
-	 		for(int j = 0; j < SIZE; j++){
-	 			if(board[i][j] == ' '){
-	 				return false;
-	 			}
-	 		}
-	 	}
-	 	return true;
-	 }
-
-	/* public static void displayBoard(){
-	 	int k = 0;
-	 	for(int i = 0; i < SIZE; i++){
-	 		for(int j = 0; j < SIZE; j++){
-	 			if(board[i][j] == ' '){
-	 				System.out.print(k);
-	 			}
-	 			else{
-	 				System.out.print(board[i][j]);
-	 			}
-	 			k++;
-	 		}
-	 		System.out.println();
-	 	}
-	 }*/
 	 
 	 public void setMove(int move, char player){
 		int row = (move - 1) / 3;
@@ -103,5 +32,7 @@ public class Board {
 	 public char[][] getBoard(){
 	 	return this.board;
 	 }
-
+	 public char getBoardValue(int row, int col){
+	 	return board[row][col];
+	 }
 }
