+++ src/test/java/is/ru/tictactoetest/BoardTest.java
@@ -16,22 +16,12 @@ public class BoardTest {
 
 	} */
 
-	@Test
-	public void testFull(){
-		
-		Board b = new Board();
-	 	assertFalse(b.full());
-	}
+	
 	
 	@Test
 	public void testSetMove() {
 		Board b = new Board();
     	b.setMove(8, 'O');
 		assertTrue(b.setMoved(8));
-	}
-
-	@Test
-	public void testWin(){
-		assertTrue(Board.win(3));
 	}
 }
